Refactor codebase to remove duplicated logic found by jscpd
- Extracted duplicate Upgrades loop from `MobileUpgrades.tsx` and `UpgradesPanel.tsx` into `UpgradesList.tsx`.
- Refactored next balls calculation in `FrameManager.tsx` to utilize a local closure callback, eliminating duplicated rendering maps.
- Created `multithread.utils.ts` and `test-utils.ts` to abstract duplicated setups across test cases.
- Ran all tests (`pnpm test:run`) successfully.
