+++ src/test/multithread.kernel.test.ts
@@ -2,42 +2,15 @@ import { describe, it, expect } from 'vitest';
 import { simulateStep } from '../engine/multithread/kernel';
 import type { SimInput } from '../engine/multithread/kernel';
 import { stepBallFrame } from '../engine/collision';
-import type { Ball, Brick } from '../store/types';
+import { mockArena, mockBrick, mockBalls } from './multithread.utils';
 
 describe('simulateStep kernel (parity with stepBallFrame)', () => {
   it('produces identical next position/velocity for a small batch', () => {
-    const arena = { width: 12, height: 10, depth: 8 };
-
-    const brick: Brick = {
-      id: 'brick-1',
-      type: 'normal',
-      position: [0, 0, 0],
-      health: 10,
-      maxHealth: 10,
-      color: '#fff',
-      value: 10,
-    };
-
-    const balls: Ball[] = [
-      {
-        id: 'ball-1',
-        position: [-0.1, 0, 0],
-        velocity: [0.2, 0, 0],
-        radius: 0.5,
-        damage: 1,
-        color: '#fff',
-      },
-      {
-        id: 'ball-2',
-        position: [1, 0.2, -0.3],
-        velocity: [-0.1, 0.05, 0.05],
-        radius: 0.25,
-        damage: 2,
-        color: '#0ff',
-      },
-    ];
-
+    const arena = mockArena;
+    const brick = mockBrick;
+    const balls = mockBalls;
     const count = balls.length;
+
     const positions = new Float32Array(count * 3);
     const velocities = new Float32Array(count * 3);
     const radii = new Float32Array(count);
